% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/powopt.R
\name{powCD}
\alias{powCD}
\title{Coordinate descent for penalized regression with power penalty}
\source{
For \eqn{q\le 1}, uses coordinate descent algorithm given by Marjanovic and Solo (2014), modified to accomodate X that do not have standardized columns. \cr

Box, G. E. P., and G. C. Tiao. "Bayesian inference in statistical inference." Adison-Wesley, Reading, Mass (1973). \cr

Marjanovic, Goran, and Victor Solo. "\eqn{l_q} Sparsity Penalized Linear Regression With Cyclic Descent." IEEE Transactions on Signal Processing 62.6 (2014): 1464-1475. \cr
}
\usage{
\code{powCD(X = X, y = y, sigma.sq = 1, lambda = 1,
      q = 1, max.iter = 10000,
      print.iter = FALSE,
      tol = 10^(-7), ridge.eps = 0,
      rand.restart = 0)}
}
\arguments{
\item{\code{X}}{design matrix}

\item{\code{y}}{response vector}

\item{\code{sigma.sq}}{scalar value \eqn{\sigma^2}}

\item{\code{lambda}}{scalar value \eqn{\lambda}}

\item{\code{q}}{scalar value \eqn{q}}

\item{\code{max.iter}}{maximum number of iterations for coordinate descent}

\item{\code{print.iter}}{logical value indicating whether iteration count for coordinate descent should be printed}

\item{\code{tol}}{scalar tolerance value for assessing convergence of objective function}

\item{\code{ridge.eps}}{ridge regression tuning parameter for obtaining starting value of \eqn{\beta} in coordinate descent, defult is zero}

\item{\code{rand.restart}}{number of times coordinate descent should be repeated from random starting value for \eqn{\beta} after an initial application of coordinate descent starting from ridge solution, needed when \eqn{X} is not orthogonal because the coordinate descent algorithm is not guaranteed to converge to the global optimum for all non-orthogonal \eqn{X} when \eqn{q \le 1}}

\item{\code{start}}{starting value, set to null by default}
}
\value{
Returns a vector of optimal values for \eqn{\beta}. If the coordinate descent algorithm does not meet the optimality conditions given in Marjanovic and Solo (2014), a vector of \code{NA}'s is returned.

Note that a non-\code{NA} solution for \eqn{\beta} for any value of \eqn{q} guarantees the global minimum has been attained when \eqn{X} is full rank and orthogonal. Otherwise, when \eqn{q \le 1} the solution will only correspond to a global minimum when the conditions on \eqn{X} given in Marjanovic and Solo (2014) are satisfied.
}
\description{
Gives the value of of the length \eqn{p} vector \eqn{\beta} that minimizes: \cr
\deqn{(y - X\beta)^2/(2\sigma^2) + \lambda ||\beta||^q_q} \cr
for fixed \eqn{y}, \eqn{X}, \eqn{\sigma^2 > 0}, \eqn{\lambda > 0 0} and \eqn{q > 0}. \cr \cr

This corresponds to finding the posterior mode for \eqn{beta} given \eqn{X}, \eqn{y}, \eqn{\sigma^2}, \eqn{\lambda} and \eqn{q} under the model:
\deqn{y = X\beta + \sigma Z}
\deqn{p(\beta_i) = q\lambda^(1/q) exp(-\lambda|\beta_i|^q)/(2 \Gamma(1/q)),}
where elements of Z are independent, standard normal random variates.

This distribution for elements of \eqn{\beta} is a generalized normal distribution
for \eqn{\beta} with scale \eqn{\alpha = \lambda^(-1/q)} and shape \eqn{\beta = q} (Box and Tiao, 1973).

For \eqn{q \le 1}, uses coordinate descent algorithm given in Marjanovic and Solo (2014), modified to accomodate X that do not have standardized columns.
}
