% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/powopt.R
\name{powThresh}
\alias{powThresh}
\title{Thresholding function for penalized regression with power penalty}
\usage{
powThresh(z, lambda, q)
}
\description{
Gives the value of beta that minimizes (beta - z)^2/2 + lambda|beta|^q for fixed z, lambda and q. For q <= 1, uses thresholding function given in Marjanovic and Solo (2014). For q > 1, computed using fixed point iteration.
}
