% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/powopt.R
\name{powObj}
\alias{powObj}
\title{Penalized likelihood function, used to differentiate between local modes}
\usage{
powObj(beta, X, y, sigma.sq, lambda, q, Q = NULL, l = NULL)
}
\description{
Penalized likelihood function, used to differentiate between local modes
}
