% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{powObj}
\alias{powObj}
\title{Penalized likelihood objective function}
\usage{
\code{powObj(beta, Q, l, sigmasq = 1, lambda = 1, q = 1)}
}
\arguments{
\item{\code{beta}}{length p vector of coefficient values to compute objective function for}

\item{\code{Q}}{p\eqn{\times}p matrix corresponding to \eqn{X'X}}

\item{\code{l}}{length p vector corresponding to \eqn{X'y}}

\item{\code{sigmasq}}{scalar noise variance}

\item{\code{lambda}}{scalar multiplier applied to penalty}

\item{\code{q}}{scalar power of penalty function}
}
\description{
Letting \eqn{Q =X'X} and \eqn{l = X'y}, gives the value: \cr
\deqn{-((\beta'Q\beta - 2\beta'l)/(2\sigma^2) + \lambda \sum_{j = 1}^p |\beta_j|^q)} \cr
for fixed \eqn{z}, \eqn{\lambda > 0} and \eqn{q > 0}.
}
